Updated Dark mode
Added a light mode and dark mode theme in the build widget which establishes overall themes so the navigation bar and buttons can also change colors when dark mode is enabled. and made the dark mode button round to match the others. and it switches to a light mode button when light mode is enabled.